requirements.txt
# Core dependencies
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=8.3.0
einops>=0.4.0

# Optional: for advanced features
# basicsr>=1.4.0  # For automatic model downloading (if needed)